pyproject.toml
[project]
name = "reflection-agent-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "dajong"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "grandalf (>=0.8,<0.9)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
